server/requirements.txt
fastapi
uvicorn[standard]
pymongo
pinecone
langchain
langchain-mongodb
langchain-google-genai
langchain-openai
langchain-pinecone
python-jose[cryptography]
python-multipart
pydantic
pydantic-settings
python-dotenv
openai
langchain_google_genai
# PDF processing libraries
PyPDF2
langchain-text-splitters
